"Refactor auth module paths for consistent directory structure."
